refactor(test): split anthropic-to-codex and tool-format tests (#627, #629)
- anthropic-to-codex.test.ts: 910 → 484 lines (system, messages, thinking, model)
- anthropic-to-codex-tools.test.ts: 462 lines (tools, tool_result, mixed content)
- tool-format.test.ts: 827 → 390 lines (core OpenAI/Anthropic/Gemini conversions)
- tool-format-edge-cases.test.ts: 404 lines (schema normalization, Read pages hint, web_search)

All 113 translation tests pass. No logic changes.
